Moved out some logic into a controller
Moved the logic to get commits, get summary, and get tumblr posts to
controller.js. This way commitlr.js can stay just for the dom manips and
stuff. "Shitty code, as you know, is a lot like gravity. All it takes is
a little push." said the Joker and we shouldn't let him win.